+++ apps/prairielearn/src/pages/instructorAssessmentManualGrading/instanceQuestion/instanceQuestion.ts
@@ -119,10 +119,6 @@ router.get(
       return null;
     });
 
-    if (instance_question == null) {
-      throw new error.HttpStatusError(404, 'Instance question not found');
-    }
-
     const aiGradingEnabled = await features.enabledFromLocals('ai-grading', res.locals);
 
     const instanceQuestionGroups = await selectInstanceQuestionGroups({
@@ -175,14 +171,28 @@ router.get(
         // try to avoid errors when extracting the explanation. Note that for some
         // time, the explanation wasn't included in the completion at all, so it
         // may legitimately be missing.
+        //
+        // The responses API changed the format here. We'll need to handle both the
+        // old and new formats.
         const explanation = run(() => {
           const completion = ai_grading_job_data.completion;
-          if (completion == null) return null;
+          if (!completion) return null;
 
-          const explanation = completion?.choices?.[0]?.message?.parsed?.explanation;
-          if (typeof explanation !== 'string') return null;
+          if (completion.choices) {
+            const explanation = completion?.choices?.[0]?.message?.parsed?.explanation;
+            if (typeof explanation !== 'string') return null;
 
-          return explanation.trim() || null;
+            return explanation.trim() || null;
+          }
+
+          if (completion.output_parsed) {
+            const explanation = completion?.output_parsed?.explanation;
+            if (typeof explanation !== 'string') return null;
+
+            return explanation.trim() || null;
+          }
+
+          return null;
         });
 
         aiGradingInfo = {
@@ -377,8 +387,7 @@ router.post(
       qs.parse(qs.stringify(req.body), { parseArrays: false }),
     );
     if (body.__action === 'add_manual_grade') {
-      req.session.skip_graded_submissions =
-        body.skip_graded_submissions ?? req.session.skip_graded_submissions ?? true;
+      req.session.skip_graded_submissions = body.skip_graded_submissions;
 
       const manual_rubric_data = res.locals.assessment_question.manual_rubric_id
         ? {
@@ -442,8 +451,7 @@ router.post(
         }),
       );
     } else if (body.__action === 'next_instance_question') {
-      req.session.skip_graded_submissions =
-        body.skip_graded_submissions ?? req.session.skip_graded_submissions ?? true;
+      req.session.skip_graded_submissions = body.skip_graded_submissions;
 
       const use_instance_question_groups = await run(async () => {
         const aiGradingMode =