+++ apps/prairielearn/src/pages/instructorCourseAdminTopics/components/EditTopicsModal.tsx
@@ -26,7 +26,7 @@ export function EditTopicsModal({
   function handleSubmit() {
     setInvalidName(!selectedTopic?.name);
     setInvalidColor(!selectedTopic?.color);
-    if (!selectedTopic?.name || !selectedTopic?.color) {
+    if (!selectedTopic?.name || !selectedTopic.color) {
       return;
     } else {
       handleModalUpdate();
@@ -60,7 +60,7 @@ export function EditTopicsModal({
                 onChange={(e) =>
                   setSelectedTopic({
                     ...selectedTopic,
-                    name: (e.target as HTMLInputElement)?.value,
+                    name: (e.target as HTMLInputElement).value,
                   })
                 }
               />
@@ -74,11 +74,11 @@ export function EditTopicsModal({
                 <select
                   class={clsx('form-select', invalidColor && 'is-invalid')}
                   id="topicColor"
-                  value={selectedTopic.color ?? 'Select a color'}
+                  value={selectedTopic.color}
                   onChange={(e) =>
                     setSelectedTopic({
                       ...selectedTopic,
-                      color: (e.target as HTMLSelectElement)?.value ?? 'gray1',
+                      color: (e.target as HTMLSelectElement).value,
                     })
                   }
                 >
@@ -101,7 +101,7 @@ export function EditTopicsModal({
                     width="32"
                     height="32"
                     style={{
-                      fill: `var(--color-${selectedTopic.color ?? 'gray1'})`,
+                      fill: `var(--color-${selectedTopic.color})`,
                       rx: 'var(--bs-border-radius)',
                       ry: 'var(--bs-border-radius)',
                     }}
@@ -122,7 +122,7 @@ export function EditTopicsModal({
                 onChange={(e) =>
                   setSelectedTopic({
                     ...selectedTopic,
-                    description: (e.target as HTMLTextAreaElement)?.value,
+                    description: (e.target as HTMLTextAreaElement).value,
                   })
                 }
               />