+++ apps/prairielearn/src/pages/instructorCourseAdminTopics/components/InstructorCourseAdminTopicsTable.tsx
@@ -13,7 +13,7 @@ const emptyTopic: Topic = {
   color: '',
   course_id: '',
   description: '',
-  id: uuidv4(),
+  id: '',
   implicit: false,
   json_comment: '',
   name: '',
@@ -64,6 +64,7 @@ export function InstructorCourseAdminTopicsTable({
     setAddTopic(true);
     setSelectedTopic({
       ...emptyTopic,
+      id: uuidv4(),
       // Pick a random initial color.
       color: ColorJsonSchema.options[Math.floor(Math.random() * ColorJsonSchema.options.length)],
     });
@@ -129,7 +130,7 @@ export function InstructorCourseAdminTopicsTable({
               {topicsState.map((topic, index) => {
                 return (
                   <tr key={topic.name}>
-                    {editMode && allowEdit ? (
+                    {editMode && allowEdit && (
                       <td class="align-middle">
                         <div class="d-flex align-items-center">
                           <button
@@ -150,8 +151,6 @@ export function InstructorCourseAdminTopicsTable({
                           </button>
                         </div>
                       </td>
-                    ) : (
-                      ''
                     )}
                     <td class="align-middle">{index + 1}</td>
                     <td class="align-middle">
@@ -166,7 +165,7 @@ export function InstructorCourseAdminTopicsTable({
               })}
               {editMode ? (
                 <tr>
-                  <td colSpan={editMode ? 6 : 4}>
+                  <td colSpan={5}>
                     <button class="btn btn-sm btn-ghost" type="button" onClick={handleNewTopic}>
                       <i class="fa fa-plus" aria-hidden="true" /> New Topic
                     </button>