+++ packages/preact/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @prairielearn/preact
 
+## 1.0.2
+
+### Patch Changes
+
+- c72a4b8: Upgrade dependencies
+- Updated dependencies [c72a4b8]
+  - @prairielearn/browser-utils@2.2.14
+  - @prairielearn/compiled-assets@3.2.7
+  - @prairielearn/error@2.0.20
+  - @prairielearn/html@4.0.20
+  - @prairielearn/preact-cjs@1.1.4
+  - @prairielearn/utils@2.0.1
+
 ## 1.0.1
 
 ### Patch Changes